Add files via upload
This folder was created when I attempted to get a Web Dev certificate. It is coded in HTML and CSS. I've decided to keep it the way it is, flaws and all. Maybe I'll look back at this code and laugh at how new I was and still am to this world.